@@ -1,149 +1,56 @@
 {
-  "name": "fuse-react-app",
-  "version": "8.1.0",
+  "name": "explorer-frontend",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc -b && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "format": "prettier --write .",
+    "preview": "vite preview"
+  },
   "dependencies": {
-    "@emotion/cache": "11.9.3",
-    "@emotion/react": "11.9.3",
-    "@emotion/styled": "11.9.3",
-    "@hookform/resolvers": "2.9.5",
-    "@mui/base": "5.0.0-alpha.89",
-    "@mui/icons-material": "5.8.4",
-    "@mui/lab": "5.0.0-alpha.90",
-    "@mui/material": "5.9.0",
-    "@mui/styles": "5.9.0",
-    "@mui/system": "5.9.0",
-    "@mui/utils": "5.9.0",
-    "@mui/x-data-grid": "5.13.0",
-    "@mui/x-date-pickers": "5.0.0-beta.0",
-    "@reduxjs/toolkit": "1.8.3",
-    "autosuggest-highlight": "3.3.1",
+    "@reduxjs/toolkit": "^2.2.6",
+    "@tailwindcss/aspect-ratio": "^0.4.2",
+    "@tailwindcss/typography": "^0.5.13",
     "axios": "^1.7.2",
-    "axios-mock-adapter": "1.21.1",
-    "body-parser": "^1.20.2",
-    "clsx": "1.2.1",
-    "crypto-js": "^4.2.0",
-    "date-fns": "2.28.0",
-    "draft-js": "0.11.7",
-    "draftjs-to-html": "0.9.1",
-    "express": "^4.19.1",
-    "framer-motion": "6.5.0",
-    "history": "5.3.0",
-    "i18next": "21.8.14",
-    "jwt-decode": "3.1.2",
-    "keycode": "2.2.1",
-    "lint-staged": "^15.2.7",
-    "lodash": "4.17.21",
+    "clsx": "^2.1.1",
+    "i18next": "^23.12.1",
+    "i18next-http-backend": "^2.5.2",
     "lottie-react": "^2.4.0",
-    "mobile-detect": "1.4.5",
-    "moment": "2.29.4",
-    "notistack": "1.0.6-next.3",
-    "perfect-scrollbar": "1.5.5",
-    "prismjs": "1.28.0",
-    "prop-types": "15.8.1",
-    "qs": "6.11.0",
     "qubic-ts-library": "^0.1.1",
-    "react": "18.2.0",
-    "react-autosuggest": "10.1.0",
-    "react-dom": "18.2.0",
-    "react-draft-wysiwyg": "1.14.7",
-    "react-hook-form": "7.33.1",
-    "react-i18next": "11.18.1",
-    "react-infinite-scroll-component": "^6.1.0",
-    "react-popper": "2.3.0",
-    "react-redux": "8.0.2",
-    "react-router-dom": "6.3.0",
-    "react-swipeable": "7.0.0",
-    "redux-logger": "4.0.0",
-    "request": "^2.88.2",
-    "styled-components": "5.3.5",
-    "stylis": "4.1.1",
-    "stylis-plugin-rtl": "2.1.1",
-    "web-vitals": "2.1.4",
-    "yup": "0.32.11"
-  },
-  "peerDependencies": {
-    "autoprefixer": "^10.4.19",
-    "postcss": "8.4.13",
-    "react": "18.2.0",
-    "react-dom": "18.2.0"
-  },
-  "resolutions": {
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "babel-loader": "8.1.0"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-i18next": "^14.1.3",
+    "react-redux": "^9.1.2",
+    "react-router-dom": "^6.24.1",
+    "redux-logger": "^3.0.6",
+    "tailwind-merge": "^2.4.0"
   },
   "devDependencies": {
-    "@babel/core": "7.17.10",
-    "@babel/eslint-parser": "7.17.0",
-    "@babel/node": "7.17.10",
-    "@babel/preset-env": "7.17.10",
-    "@babel/preset-react": "7.16.7",
-    "@tailwindcss/aspect-ratio": "^0.4.0",
-    "@tailwindcss/line-clamp": "^0.4.0",
-    "@tailwindcss/typography": "^0.5.4",
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
-    "ajv": "^8.16.0",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@typescript-eslint/eslint-plugin": "5.62.0",
+    "@typescript-eslint/parser": "5.62.0",
+    "@vitejs/plugin-react": "^4.3.1",
     "autoprefixer": "^10.4.19",
-    "concurrently": "^8.2.2",
-    "cross-env": "7.0.3",
-    "eslint": "7.32.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-config-react-app": "^6.0.0",
-    "eslint-plugin-flowtype": "^5.10.0",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.26.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-unused-imports": "^1.1.5",
-    "husky": "^9.0.11",
-    "immutable": "4.0.0",
-    "postcss": "8.4.5",
-    "prettier": "^2.4.1",
-    "promise": "8.1.0",
-    "react-app-alias": "2.2.0",
-    "react-app-rewired": "2.2.1",
-    "react-scripts": "^5.0.1",
-    "source-map-explorer": "2.5.2",
-    "tailwindcss": "3.1.6",
-    "typescript": "4.6.4",
-    "util": "^0.12.4"
-  },
-  "scripts": {
-    "start": "react-app-rewired  start",
-    "start-api-proxy": "concurrently \"node dev-proxy-qli.js\" \"node dev-proxy-arc.js\"",
-    "build": "cross-env GENERATE_SOURCEMAP=false react-app-rewired build",
-    "test": "react-app-rewired test --env=node",
-    "eject": "react-app-rewired eject",
-    "build-docs": "babel-node --presets @babel/preset-env src/app/main/documentation/material-ui-components/build.js",
-    "analyze": "react-app-rewired build && source-map-explorer 'build/static/js/*.js' --html analyze-result.html",
-    "lint": "eslint . --fix",
-    "format": "prettier --write .",
-    "prepare": "husky"
-  },
-  "engines": {
-    "node": ">=14.0.0",
-    "npm": ">=7.0.0"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 3 safari version"
-    ]
-  },
-  "lint-staged": {
-    "*.js": [
-      "prettier --write",
-      "eslint --fix"
-    ]
+    "eslint": "8.56.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-typescript": "17.0.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsx-a11y": "^6.9.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-react": "^7.34.3",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-plugin-react-refresh": "^0.4.7",
+    "eslint-plugin-unused-imports": "2.0.0",
+    "postcss": "^8.4.39",
+    "prettier": "3.3.2",
+    "tailwindcss": "^3.4.4",
+    "typescript": "^5.5.3",
+    "vite": "^5.3.3",
+    "vite-plugin-svgr": "^4.2.0"
   }
 }
